requirements.txt
argparse
cssselect
pyquery==1.2.9
numpy
biopython>=1.65
helperlibs
pysvg
pyExcelerator
backports.lzma
bcbio-gff
ete2
networkx
pandas
matplotlib
scipy
scikit-learn
